comments, clean up, move components
