+++ frontend/src/components/views/index.ts
@@ -2,8 +2,6 @@ import { PreGameHostView } from "./pre-game/pre-game-host-view";
 import { PreGamePlayerView } from "./pre-game/pre-game-player-view";
 import { PickingDrawerView } from "./playing/picking/picking-drawer-view";
 import { PickingGuesserView } from "./playing/picking/picking-guesser-view";
-import { PostGameHostView } from "./post-game/post-game-host-view";
-import { PostGamePlayerView } from "./post-game/post-game-player-view";
 import { DrawingDrawerView } from "./playing/drawing/drawing-drawer-view";
 import { DrawingGuesserView } from "./playing/drawing/drawing-guesser-view";
 import { PostDrawingView } from "./playing/post-drawing-view";
@@ -20,6 +18,4 @@ export {
 	DrawingDrawerView,
 	DrawingGuesserView,
 	PostDrawingView,
-	PostGameHostView,
-	PostGamePlayerView,
 };
