--- frontend/src/components/views/post-game/post-game-player-view.tsx
@@ -1,7 +1,0 @@
-export function PostGamePlayerView() {
-	return (
-		<div className="flex h-full flex-col items-center justify-center">
-			<div>PostGamePlayerView</div>
-		</div>
-	);
-}
