+++ frontend/src/lib/realtime.ts
@@ -5,11 +5,15 @@ export const getRealtimeHref = () => {
 	const protocol = isDev ? "ws" : "wss";
 	const isBranch = import.meta.env.BRANCH_SOCKET_HOST;
 
-	const host = isDev
-		? "localhost:8080"
-		: isBranch
-		? import.meta.env.BRANCH_SOCKET_HOST
-		: "realtime.sketchwithfriends.com";
+	let host;
+	if (isDev) {
+		host = import.meta.env.VITE_SOCKET_HOST;
+	} else if (isBranch) {
+		host = import.meta.env.BRANCH_SOCKET_HOST;
+	} else {
+		host = "realtime.sketchwithfriends.com";
+	}
+
 	return `${protocol}://${host}`;
 };
